refactor: 유효성 검사 공통 로직 훅 생성
